Apply refactor to patient
