+++ packages/shared-src/src/models/fhir/Patient/getQueryOptions.js
@@ -1,0 +1,17 @@
+export function getQueryOptions(models) {
+  const { Patient, PatientAdditionalData } = models;
+
+  const patientOptions = {
+    include: [
+      {
+        model: PatientAdditionalData,
+        as: 'additionalData',
+        limit: 1,
+      },
+    ],
+  };
+
+  return {
+    [Patient.tableName]: patientOptions,
+  };
+}
